refactors amortizing calculation

+++ src/main/java/model/Loan.java
@@ -1,15 +1,16 @@
 package main.java.model;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 
 public class Loan {
 
     private LoanType loanType;
     private Rate[] rates;
-    private int total;
+    private BigDecimal total;
     private Optional<Double> monthlyRate;
 
-    public Loan(LoanType loanType, Rate[] rates, int total, Optional<Double> monthlyRate) {
+    public Loan(LoanType loanType, Rate[] rates, BigDecimal total, Optional<Double> monthlyRate) {
         this.loanType = loanType;
         this.rates = rates;
         this.total = total;
@@ -24,7 +25,7 @@ public class Loan {
         return rates;
     }
 
-    public int getTotal() { return total; }
+    public BigDecimal getTotal() { return total; }
 
     public Optional<Double> getMonthlyRate() { return monthlyRate; }
 }
